+++ .github/workflows/build.yml
@@ -78,14 +78,12 @@ jobs:
             --ignore=tests/reaxff_test.py \
             --ignore=tests/tpu_test.py \
             --ignore=tests/uma_test.py \
-            --ignore=tests/test_uma_comparison.py \
             --ignore=tests/uma_100_structures_test.py \
             --ignore=tests/uma_pretrained_test.py \
             --ignore=tests/uma_relaxation_test.py
         elif [ "${{ matrix.test-suite }}" == "uma" ]; then
           JAX_ENABLE_X64=1 uv run pytest \
             tests/uma_test.py \
-            tests/test_uma_comparison.py \
             tests/uma_pretrained_test.py \
             tests/uma_relaxation_test.py \
             tests/uma_100_structures_test.py \
