--- tests/test_uma_comparison.py
@@ -1,1523 +1,0 @@
-"""
-Standalone test script for comparing JAX UMA implementation with PyTorch.
-
-This script:
-1. Creates both PyTorch and JAX UMA models with identical configurations
-2. Transfers weights from PyTorch to JAX
-3. Runs forward passes on both models with identical inputs
-4. Compares outputs to verify numerical equivalence
-
-Usage:
-    python test_uma_comparison.py
-"""
-
-from __future__ import annotations
-
-import math
-import os
-import sys
-from dataclasses import dataclass, field
-from typing import Dict, List, Literal, Optional, Tuple, NamedTuple
-
-import pytest
-import numpy as np
-
-pytest.importorskip('torch', reason='torch required for UMA comparison tests')
-
-import jax
-import jax.numpy as jnp
-from jax.nn import initializers
-import flax.linen as nn
-
-EPS = 1e-7
-
-
-# -----------------------------------------------------------------------------
-# Common utilities: SO(3) coefficient mapping
-# -----------------------------------------------------------------------------
-
-
-class CoefficientMapping(NamedTuple):
-  """Helper class for coefficients used to reshape l <--> m ordering."""
-
-  lmax: int
-  mmax: int
-  to_m: jnp.ndarray
-  l_harmonic: jnp.ndarray
-  m_harmonic: jnp.ndarray
-  m_complex: jnp.ndarray
-  m_size: Tuple[int, ...]
-  res_size: int
-  coefficient_idx_cache: Dict[Tuple[int, int], jnp.ndarray]
-
-
-def _complex_idx(
-  m: int, lmax: int, mmax: int, m_complex: jnp.ndarray, l_harmonic: jnp.ndarray
-) -> Tuple[List[int], List[int]]:
-  """Get indices for real and imaginary parts of order m coefficients."""
-  indices = np.arange(len(l_harmonic))
-  m_complex_np = np.array(m_complex)
-  l_harmonic_np = np.array(l_harmonic)
-
-  mask_r = (l_harmonic_np <= lmax) & (m_complex_np == m)
-  mask_idx_r = indices[mask_r].tolist()
-
-  mask_idx_i = []
-  if m != 0:
-    mask_i = (l_harmonic_np <= lmax) & (m_complex_np == -m)
-    mask_idx_i = indices[mask_i].tolist()
-
-  return mask_idx_r, mask_idx_i
-
-
-def create_coefficient_mapping(lmax: int, mmax: int) -> CoefficientMapping:
-  """Create a CoefficientMapping instance."""
-  l_harmonic_list = []
-  m_harmonic_list = []
-  m_complex_list = []
-
-  for l in range(lmax + 1):
-    mmax_l = min(mmax, l)
-    for m in range(-mmax_l, mmax_l + 1):
-      m_complex_list.append(m)
-      m_harmonic_list.append(abs(m))
-      l_harmonic_list.append(l)
-
-  l_harmonic = jnp.array(l_harmonic_list, dtype=jnp.int32)
-  m_harmonic = jnp.array(m_harmonic_list, dtype=jnp.int32)
-  m_complex = jnp.array(m_complex_list, dtype=jnp.int32)
-  res_size = len(l_harmonic_list)
-
-  num_coefficients = res_size
-  to_m = np.zeros([num_coefficients, num_coefficients])
-  m_size_list = [0] * (mmax + 1)
-
-  offset = 0
-  for m in range(mmax + 1):
-    idx_r, idx_i = _complex_idx(m, lmax, mmax, m_complex, l_harmonic)
-
-    for idx_out, idx_in in enumerate(idx_r):
-      to_m[idx_out + offset, idx_in] = 1.0
-    offset = offset + len(idx_r)
-
-    for idx_out, idx_in in enumerate(idx_i):
-      to_m[idx_out + offset, idx_in] = 1.0
-    offset = offset + len(idx_i)
-
-    # m_size includes both real and imaginary parts
-    m_size_list[m] = len(idx_r) + len(idx_i)
-
-  to_m = jnp.array(to_m)
-
-  coefficient_idx_cache = {}
-  for l in range(lmax + 1):
-    for m_val in range(lmax + 1):
-      mask = (l_harmonic <= l) & (m_harmonic <= m_val)
-      indices = jnp.arange(len(mask))
-      mask_indices = jnp.where(mask, indices, -1)
-      mask_indices = mask_indices[mask_indices >= 0]
-      coefficient_idx_cache[(l, m_val)] = mask_indices
-
-  return CoefficientMapping(
-    lmax=lmax,
-    mmax=mmax,
-    to_m=to_m,
-    l_harmonic=l_harmonic,
-    m_harmonic=m_harmonic,
-    m_complex=m_complex,
-    m_size=tuple(m_size_list),
-    res_size=res_size,
-    coefficient_idx_cache=coefficient_idx_cache,
-  )
-
-
-def coefficient_idx(
-  mapping: CoefficientMapping, lmax: int, mmax: int
-) -> jnp.ndarray:
-  """Get indices of coefficients with degree <= lmax and order <= mmax."""
-  if (lmax, mmax) in mapping.coefficient_idx_cache:
-    return mapping.coefficient_idx_cache[(lmax, mmax)]
-  mask = (mapping.l_harmonic <= lmax) & (mapping.m_harmonic <= mmax)
-  indices = jnp.arange(len(mask))
-  return indices[mask]
-
-
-# -----------------------------------------------------------------------------
-# Common utilities: SO(3) grid
-# -----------------------------------------------------------------------------
-
-
-class SO3Grid(NamedTuple):
-  """Helper class for grid representation of spherical harmonic irreps."""
-
-  lmax: int
-  mmax: int
-  lat_resolution: int
-  long_resolution: int
-  mapping: CoefficientMapping
-  to_grid_mat: jnp.ndarray
-  from_grid_mat: jnp.ndarray
-  rescale: bool
-
-
-def _associated_legendre(l: int, m: int, x: np.ndarray) -> np.ndarray:
-  """Compute associated Legendre polynomial P_l^m(x)."""
-  if m > l:
-    return np.zeros_like(x)
-
-  pmm = np.ones_like(x)
-  if m > 0:
-    somx2 = np.sqrt((1 - x) * (1 + x))
-    fact = 1.0
-    for i in range(1, m + 1):
-      pmm = -pmm * fact * somx2
-      fact += 2.0
-
-  if l == m:
-    return pmm
-
-  pmmp1 = x * (2 * m + 1) * pmm
-  if l == m + 1:
-    return pmmp1
-
-  pll = np.zeros_like(x)
-  for ll in range(m + 2, l + 1):
-    pll = ((2 * ll - 1) * x * pmmp1 - (ll + m - 1) * pmm) / (ll - m)
-    pmm = pmmp1
-    pmmp1 = pll
-
-  return pll
-
-
-def _compute_grid_matrices(
-  lmax: int,
-  mmax: int,
-  lat_resolution: int,
-  long_resolution: int,
-  rescale: bool,
-  mapping: CoefficientMapping,
-) -> Tuple[jnp.ndarray, jnp.ndarray]:
-  """Compute transformation matrices between coefficients and grid."""
-  theta = np.linspace(0, np.pi, lat_resolution, endpoint=False)
-  theta = theta + np.pi / (2 * lat_resolution)
-  phi = np.linspace(0, 2 * np.pi, long_resolution, endpoint=False)
-
-  sph_size = (lmax + 1) ** 2
-  to_grid = np.zeros((lat_resolution, long_resolution, sph_size))
-  from_grid = np.zeros((lat_resolution, long_resolution, sph_size))
-
-  cos_theta = np.cos(theta)
-
-  for l in range(lmax + 1):
-    for m in range(-l, l + 1):
-      idx = l * l + l + m
-      plm = _associated_legendre(l, abs(m), cos_theta)
-      norm = np.sqrt(
-        (2 * l + 1)
-        / (4 * np.pi)
-        * math.factorial(l - abs(m))
-        / math.factorial(l + abs(m))
-      )
-      if m >= 0:
-        azimuth = np.cos(m * phi)
-      else:
-        azimuth = np.sin(abs(m) * phi)
-
-      for i, t in enumerate(theta):
-        for j, p in enumerate(phi):
-          to_grid[i, j, idx] = norm * plm[i] * azimuth[j]
-          from_grid[i, j, idx] = norm * plm[i] * azimuth[j]
-
-  sin_theta = np.sin(theta)
-  quad_weight = 2 * np.pi / long_resolution * np.pi / lat_resolution
-  for i in range(lat_resolution):
-    from_grid[i, :, :] *= sin_theta[i] * quad_weight
-
-  if rescale and lmax != mmax:
-    for lval in range(lmax + 1):
-      if lval <= mmax:
-        continue
-      start_idx = lval**2
-      length = 2 * lval + 1
-      rescale_factor = np.sqrt(length / (2 * mmax + 1))
-      to_grid[:, :, start_idx : start_idx + length] *= rescale_factor
-      from_grid[:, :, start_idx : start_idx + length] *= rescale_factor
-
-  coef_idx = coefficient_idx(mapping, lmax, mmax)
-  coef_idx_np = np.array(coef_idx)
-  to_grid = to_grid[:, :, coef_idx_np]
-  from_grid = from_grid[:, :, coef_idx_np]
-
-  return jnp.array(to_grid), jnp.array(from_grid)
-
-
-def create_so3_grid(
-  lmax: int,
-  mmax: int,
-  resolution: int | None = None,
-  rescale: bool = True,
-) -> SO3Grid:
-  """Create an SO3Grid instance for grid-based spherical harmonic operations."""
-  lat_resolution = 2 * (lmax + 1)
-  if lmax == mmax:
-    long_resolution = 2 * (mmax + 1) + 1
-  else:
-    long_resolution = 2 * mmax + 1
-
-  if resolution is not None:
-    lat_resolution = resolution
-    long_resolution = resolution
-
-  mapping = create_coefficient_mapping(lmax, lmax)
-  to_grid_mat, from_grid_mat = _compute_grid_matrices(
-    lmax, mmax, lat_resolution, long_resolution, rescale, mapping
-  )
-
-  return SO3Grid(
-    lmax=lmax,
-    mmax=mmax,
-    lat_resolution=lat_resolution,
-    long_resolution=long_resolution,
-    mapping=mapping,
-    to_grid_mat=to_grid_mat,
-    from_grid_mat=from_grid_mat,
-    rescale=rescale,
-  )
-
-
-# -----------------------------------------------------------------------------
-# Common utilities: Rotation / Wigner D-matrix
-# -----------------------------------------------------------------------------
-
-
-def safe_acos(x: jnp.ndarray) -> jnp.ndarray:
-  """Numerically stable arccos with gradient clipping."""
-  x_clamped = jnp.clip(x, -1 + EPS, 1 - EPS)
-  return jnp.arccos(x_clamped)
-
-
-def safe_atan2(y: jnp.ndarray, x: jnp.ndarray) -> jnp.ndarray:
-  """Numerically stable atan2."""
-  return jnp.arctan2(y, x)
-
-
-def init_edge_rot_euler_angles(
-  edge_distance_vec: jnp.ndarray,
-) -> Tuple[jnp.ndarray, jnp.ndarray, jnp.ndarray]:
-  """Compute Euler angles for rotating to edge-aligned frame."""
-  norm = jnp.linalg.norm(edge_distance_vec, axis=-1, keepdims=True)
-  xyz = edge_distance_vec / jnp.maximum(norm, EPS)
-  xyz = jnp.clip(xyz, -1.0, 1.0)
-
-  x, y, z = xyz[:, 0], xyz[:, 1], xyz[:, 2]
-  beta = safe_acos(y)
-  alpha = safe_atan2(x, z)
-  gamma = jnp.zeros_like(alpha)
-
-  return -gamma, -beta, -alpha
-
-
-def _compute_wigner_d_small(l: int, beta: float) -> np.ndarray:
-  """Compute small Wigner d-matrix d^l(beta)."""
-  if l == 0:
-    return np.array([[1.0]])
-
-  size = 2 * l + 1
-  d = np.zeros((size, size))
-
-  c = np.cos(beta / 2)
-  s = np.sin(beta / 2)
-
-  for m in range(-l, l + 1):
-    for mp in range(-l, l + 1):
-      idx_m = m + l
-      idx_mp = mp + l
-
-      val = 0.0
-      s_min = max(0, m - mp)
-      s_max = min(l + m, l - mp)
-
-      for s_idx in range(s_min, s_max + 1):
-        sign = (-1) ** (mp - m + s_idx)
-        num = (
-          math.factorial(l + m)
-          * math.factorial(l - m)
-          * math.factorial(l + mp)
-          * math.factorial(l - mp)
-        )
-        denom = (
-          math.factorial(l + m - s_idx)
-          * math.factorial(s_idx)
-          * math.factorial(mp - m + s_idx)
-          * math.factorial(l - mp - s_idx)
-        )
-        power_c = 2 * l + m - mp - 2 * s_idx
-        power_s = mp - m + 2 * s_idx
-
-        term = sign * np.sqrt(num) / denom
-        if power_c >= 0 and power_s >= 0:
-          term *= c**power_c * s**power_s
-        else:
-          term = 0.0
-        val += term
-
-      d[idx_m, idx_mp] = val
-
-  return d
-
-
-def compute_jacobi_matrices(lmax: int) -> List[jnp.ndarray]:
-  """Compute Jacobi matrices for Wigner D-matrix computation."""
-  Jd_list = []
-  for l in range(lmax + 1):
-    J = _compute_wigner_d_small(l, np.pi / 2)
-    Jd_list.append(jnp.array(J))
-  return Jd_list
-
-
-def load_jacobi_matrices_from_file(lmax: int) -> List[jnp.ndarray]:
-  """Load precomputed Jacobi matrices from the bundled Jd.pt file."""
-  try:
-    import torch
-
-    _JD_FILE = os.path.join(os.path.dirname(__file__), 'Jd.pt')
-    if os.path.exists(_JD_FILE):
-      Jd_torch = torch.load(_JD_FILE, map_location='cpu', weights_only=False)
-      return [jnp.array(Jd_torch[l].numpy()) for l in range(lmax + 1)]
-  except (ImportError, FileNotFoundError):
-    pass
-  return compute_jacobi_matrices(lmax)
-
-
-def _z_rot_mat(angle: jnp.ndarray, l: int) -> jnp.ndarray:
-  """Compute z-rotation matrix for degree l representation."""
-  batch_size = angle.shape[0]
-  size = 2 * l + 1
-
-  M = jnp.zeros((batch_size, size, size))
-
-  for i in range(size):
-    m = l - i
-    freq = m
-    M = M.at[:, i, i].set(jnp.cos(freq * angle))
-    j = size - 1 - i
-    M = M.at[:, i, j].set(jnp.sin(freq * angle))
-
-  return M
-
-
-def wigner_D(
-  l: int,
-  alpha: jnp.ndarray,
-  beta: jnp.ndarray,
-  gamma: jnp.ndarray,
-  Jd: jnp.ndarray,
-) -> jnp.ndarray:
-  """Compute Wigner D-matrix for degree l."""
-  Xa = _z_rot_mat(alpha, l)
-  Xb = _z_rot_mat(beta, l)
-  Xc = _z_rot_mat(gamma, l)
-
-  result = jnp.einsum('bij,jk->bik', Xa, Jd)
-  result = jnp.einsum('bij,bjk->bik', result, Xb)
-  result = jnp.einsum('bij,jk->bik', result, Jd)
-  result = jnp.einsum('bij,bjk->bik', result, Xc)
-
-  return result
-
-
-def eulers_to_wigner(
-  eulers: Tuple[jnp.ndarray, jnp.ndarray, jnp.ndarray],
-  start_lmax: int,
-  end_lmax: int,
-  Jd_list: List[jnp.ndarray],
-) -> jnp.ndarray:
-  """Compute block-diagonal Wigner D-matrix from Euler angles."""
-  alpha, beta, gamma = eulers
-  num_edges = alpha.shape[0]
-
-  size = (end_lmax + 1) ** 2 - start_lmax**2
-  wigner = jnp.zeros((num_edges, size, size))
-
-  start = 0
-  for l in range(start_lmax, end_lmax + 1):
-    block = wigner_D(l, alpha, beta, gamma, Jd_list[l])
-    block_size = 2 * l + 1
-    end = start + block_size
-    wigner = wigner.at[:, start:end, start:end].set(block)
-    start = end
-
-  return wigner
-
-
-# -----------------------------------------------------------------------------
-# Neural network layers: Radial
-# -----------------------------------------------------------------------------
-
-
-class GaussianSmearing(nn.Module):
-  """Gaussian smearing for distance expansion."""
-
-  start: float = 0.0
-  stop: float = 5.0
-  num_gaussians: int = 50
-  basis_width_scalar: float = 1.0
-
-  @nn.compact
-  def __call__(self, distances: jnp.ndarray) -> jnp.ndarray:
-    offset = jnp.linspace(self.start, self.stop, self.num_gaussians)
-    coeff = -0.5 / (self.basis_width_scalar * (offset[1] - offset[0])) ** 2
-    diff = distances[..., None] - offset
-    return jnp.exp(coeff * diff**2)
-
-
-class PolynomialEnvelope(nn.Module):
-  """Polynomial envelope function for smooth distance cutoff."""
-
-  exponent: int = 5
-
-  @nn.compact
-  def __call__(self, x: jnp.ndarray) -> jnp.ndarray:
-    p = self.exponent + 1
-    a = -(p + 1) * (p + 2) / 2
-    b = p * (p + 2)
-    c = -p * (p + 1) / 2
-    x_pow_p = x**p
-    envelope = 1.0 / x + a * x_pow_p + b * x_pow_p * x + c * x_pow_p * x * x
-    return jnp.where(x < 1.0, envelope, jnp.zeros_like(x))
-
-
-# -----------------------------------------------------------------------------
-# Neural network layers: Embeddings
-# -----------------------------------------------------------------------------
-
-
-class ChgSpinEmbedding(nn.Module):
-  """Charge/Spin embedding module."""
-
-  embedding_type: str = 'pos_emb'
-  embedding_target: str = 'charge'
-  embedding_size: int = 128
-  trainable: bool = False
-  max_val: int = 10
-
-  @nn.compact
-  def __call__(self, values: jnp.ndarray) -> jnp.ndarray:
-    if self.embedding_type == 'pos_emb':
-      return self._pos_embedding(values)
-    elif self.embedding_type == 'lin_emb':
-      return self._linear_embedding(values)
-    else:
-      raise ValueError(f'Unknown embedding type: {self.embedding_type}')
-
-  def _pos_embedding(self, values: jnp.ndarray) -> jnp.ndarray:
-    d_model = self.embedding_size
-    position = values.astype(jnp.float32)[:, None]
-    div_term = jnp.exp(jnp.arange(0, d_model, 2) * (-np.log(10000.0) / d_model))
-    pe = jnp.zeros((values.shape[0], d_model))
-    pe = pe.at[:, 0::2].set(jnp.sin(position * div_term))
-    pe = pe.at[:, 1::2].set(jnp.cos(position * div_term))
-    return pe
-
-  def _linear_embedding(self, values: jnp.ndarray) -> jnp.ndarray:
-    emb = nn.Embed(
-      num_embeddings=2 * self.max_val + 1,
-      features=self.embedding_size,
-    )
-    indices = values.astype(jnp.int32) + self.max_val
-    return emb(indices)
-
-
-class DatasetEmbedding(nn.Module):
-  """Dataset embedding module."""
-
-  embedding_size: int
-  dataset_list: Tuple[str, ...]  # Use tuple instead of list for immutability
-  trainable: bool = False
-
-  @nn.compact
-  def __call__(self, dataset_idx) -> jnp.ndarray:
-    indices = jnp.asarray(dataset_idx, dtype=jnp.int32)
-    emb = nn.Embed(
-      num_embeddings=len(self.dataset_list),
-      features=self.embedding_size,
-    )
-    return emb(indices)
-
-
-class EdgeDegreeEmbedding(nn.Module):
-  """Edge degree embedding using SO(2) convolutions."""
-
-  sphere_channels: int
-  lmax: int
-  mmax: int
-  edge_channels_list: List[int]
-  m_size: Tuple[int, ...]
-  rescale_factor: float = 5.0
-
-  @nn.compact
-  def __call__(
-    self,
-    x: jnp.ndarray,
-    x_edge: jnp.ndarray,
-    edge_index: jnp.ndarray,
-    wigner_and_M_mapping_inv: jnp.ndarray,
-    edge_envelope: jnp.ndarray,
-    node_offset: int = 0,
-  ) -> jnp.ndarray:
-    num_edges = edge_index.shape[1]
-    num_m_coeffs = sum(self.m_size)
-
-    # MLP for edge features -> scalar embedding
-    x_edge_mlp = nn.Sequential(
-      [
-        nn.Dense(self.edge_channels_list[1]),
-        nn.silu,
-        nn.Dense(self.edge_channels_list[2]),
-        nn.silu,
-        nn.Dense(num_m_coeffs * self.sphere_channels),
-      ]
-    )
-    edge_emb = x_edge_mlp(x_edge)
-    edge_emb = edge_emb.reshape(num_edges, num_m_coeffs, self.sphere_channels)
-
-    # Apply envelope
-    edge_emb = edge_emb * edge_envelope
-
-    # Rotate back to global frame
-    edge_emb = jnp.einsum('njm,nmc->njc', wigner_and_M_mapping_inv, edge_emb)
-
-    # Aggregate onto target nodes
-    target_indices = edge_index[1] - node_offset
-    new_embedding = jax.ops.segment_sum(
-      edge_emb,
-      target_indices,
-      num_segments=x.shape[0],
-    )
-
-    # Rescale and add to input
-    new_embedding = new_embedding / self.rescale_factor
-    return x + new_embedding
-
-
-# -----------------------------------------------------------------------------
-# Neural network layers: Normalization
-# -----------------------------------------------------------------------------
-
-
-class EquivariantRMSNormSH(nn.Module):
-  """Equivariant RMS normalization respecting spherical harmonic structure."""
-
-  lmax: int
-  num_channels: int
-  eps: float = 1e-6
-  affine: bool = True
-
-  @nn.compact
-  def __call__(self, x: jnp.ndarray) -> jnp.ndarray:
-    rms_per_degree = []
-    for l in range(self.lmax + 1):
-      start_idx = l * l
-      end_idx = (l + 1) * (l + 1)
-      x_l = x[:, start_idx:end_idx, :]
-      rms_l = jnp.sqrt(jnp.mean(x_l**2, axis=(1, 2), keepdims=True) + self.eps)
-      rms_per_degree.append(rms_l)
-
-    x_normalized = jnp.zeros_like(x)
-    for l in range(self.lmax + 1):
-      start_idx = l * l
-      end_idx = (l + 1) * (l + 1)
-      x_normalized = x_normalized.at[:, start_idx:end_idx, :].set(
-        x[:, start_idx:end_idx, :] / rms_per_degree[l]
-      )
-
-    if self.affine:
-      scale = self.param(
-        'scale',
-        initializers.ones,
-        (self.lmax + 1, self.num_channels),
-      )
-      for l in range(self.lmax + 1):
-        start_idx = l * l
-        end_idx = (l + 1) * (l + 1)
-        x_normalized = x_normalized.at[:, start_idx:end_idx, :].set(
-          x_normalized[:, start_idx:end_idx, :] * scale[l : l + 1, :]
-        )
-
-    return x_normalized
-
-
-def get_normalization_layer(norm_type: str, lmax: int, num_channels: int):
-  """Get normalization layer based on type."""
-  if norm_type == 'rms_norm_sh':
-    return EquivariantRMSNormSH(lmax=lmax, num_channels=num_channels)
-  else:
-    return EquivariantRMSNormSH(lmax=lmax, num_channels=num_channels)
-
-
-# -----------------------------------------------------------------------------
-# Neural network layers: SO(3) Linear
-# -----------------------------------------------------------------------------
-
-
-class SO3Linear(nn.Module):
-  """SO(3) equivariant linear layer."""
-
-  out_features: int
-  lmax: int
-  bias: bool = False
-
-  @nn.compact
-  def __call__(self, x: jnp.ndarray) -> jnp.ndarray:
-    in_features = x.shape[-1]
-    weights = []
-    for l in range(self.lmax + 1):
-      w = self.param(
-        f'weight_{l}',
-        initializers.xavier_uniform(),
-        (in_features, self.out_features),
-      )
-      weights.append(w)
-
-    out = jnp.zeros((*x.shape[:-1], self.out_features))
-    for l in range(self.lmax + 1):
-      start_idx = l * l
-      end_idx = (l + 1) * (l + 1)
-      x_l = x[:, start_idx:end_idx, :]
-      out_l = jnp.einsum('bij,jk->bik', x_l, weights[l])
-      out = out.at[:, start_idx:end_idx, :].set(out_l)
-
-    if self.bias:
-      b = self.param('bias', initializers.zeros, (self.out_features,))
-      out = out.at[:, 0, :].add(b)
-
-    return out
-
-
-# -----------------------------------------------------------------------------
-# Neural network layers: SO(2) Convolution
-# -----------------------------------------------------------------------------
-
-
-class SO2Convolution(nn.Module):
-  """SO(2) equivariant convolution."""
-
-  sphere_channels: int
-  m_output_channels: int
-  lmax: int
-  mmax: int
-  m_size: Tuple[int, ...]
-  internal_weights: bool = True
-  edge_channels_list: List[int] | None = None
-  extra_m0_output_channels: int | None = None
-
-  @nn.compact
-  def __call__(
-    self,
-    x: jnp.ndarray,
-    x_edge: jnp.ndarray | None = None,
-  ) -> Tuple[jnp.ndarray, jnp.ndarray | None]:
-    num_edges = x.shape[0]
-    num_m_coeffs = sum(self.m_size)
-
-    if self.internal_weights:
-      # Simple linear transformation per m
-      out = jnp.zeros((num_edges, num_m_coeffs, self.m_output_channels))
-      offset = 0
-      for m, m_sz in enumerate(self.m_size):
-        w = self.param(
-          f'weight_m{m}',
-          initializers.xavier_uniform(),
-          (self.sphere_channels, self.m_output_channels),
-        )
-        x_m = x[:, offset : offset + m_sz, :]
-        out_m = jnp.einsum('bmc,cd->bmd', x_m, w)
-        out = out.at[:, offset : offset + m_sz, :].set(out_m)
-        offset += m_sz
-      return out, None
-    else:
-      # Use edge features to generate weights
-      total_weights = (
-        num_m_coeffs * self.sphere_channels * self.m_output_channels
-      )
-      if self.extra_m0_output_channels is not None:
-        total_weights += (
-          self.m_size[0] * self.sphere_channels * self.extra_m0_output_channels
-        )
-
-      edge_mlp = nn.Sequential(
-        [
-          nn.Dense(self.edge_channels_list[1]),
-          nn.silu,
-          nn.Dense(self.edge_channels_list[2]),
-          nn.silu,
-          nn.Dense(total_weights),
-        ]
-      )
-      edge_weights = edge_mlp(x_edge)
-
-      # Split weights
-      main_size = num_m_coeffs * self.sphere_channels * self.m_output_channels
-      main_weights = edge_weights[:, :main_size].reshape(
-        num_edges, num_m_coeffs, self.sphere_channels, self.m_output_channels
-      )
-
-      out = jnp.einsum('bmcd,bmc->bmd', main_weights, x)
-
-      x_0_gating = None
-      if self.extra_m0_output_channels is not None:
-        extra_weights = edge_weights[:, main_size:].reshape(
-          num_edges,
-          self.m_size[0],
-          self.sphere_channels,
-          self.extra_m0_output_channels,
-        )
-        x_0 = x[:, : self.m_size[0], :]
-        # Sum over m dimension to get [batch, extra_m0_output_channels]
-        x_0_gating = jnp.einsum('bmcd,bmc->bd', extra_weights, x_0)
-
-      return out, x_0_gating
-
-
-# -----------------------------------------------------------------------------
-# Neural network layers: Activations
-# -----------------------------------------------------------------------------
-
-
-class GateActivation(nn.Module):
-  """Gated activation for equivariant features."""
-
-  lmax: int
-  mmax: int
-  num_channels: int
-  m_prime: bool = False
-
-  @nn.compact
-  def __call__(
-    self,
-    gating_scalars: jnp.ndarray,
-    x: jnp.ndarray,
-  ) -> jnp.ndarray:
-    # gating_scalars: [batch, lmax * num_channels]
-    # x: [batch, num_m_coeffs, num_channels]
-    gates = nn.sigmoid(gating_scalars)
-    gates = gates.reshape(-1, self.lmax, self.num_channels)
-
-    out = x.copy()
-    offset = 1  # Skip l=0
-    for l in range(1, self.lmax + 1):
-      mmax_l = min(self.mmax, l)
-      m_size = 2 * mmax_l + 1
-      gate_l = gates[:, l - 1 : l, :]
-      out = out.at[:, offset : offset + m_size, :].set(
-        x[:, offset : offset + m_size, :] * gate_l
-      )
-      offset += m_size
-
-    # Apply SiLU to l=0 component
-    out = out.at[:, 0:1, :].set(nn.silu(x[:, 0:1, :]))
-
-    return out
-
-
-class SeparableS2Activation(nn.Module):
-  """S2 grid-based activation."""
-
-  lmax: int
-  mmax: int
-  to_grid_mat: jnp.ndarray
-  from_grid_mat: jnp.ndarray
-
-  @nn.compact
-  def __call__(
-    self,
-    x_0_gating: jnp.ndarray,
-    x: jnp.ndarray,
-  ) -> jnp.ndarray:
-    x_grid = jnp.einsum('bai,nic->nbac', self.to_grid_mat, x)
-    x_grid = nn.silu(x_grid)
-    x = jnp.einsum('bai,nbac->nic', self.from_grid_mat, x_grid)
-    return x
-
-
-# -----------------------------------------------------------------------------
-# Model blocks
-# -----------------------------------------------------------------------------
-
-
-class Edgewise(nn.Module):
-  """Edgewise message passing module."""
-
-  sphere_channels: int
-  hidden_channels: int
-  lmax: int
-  mmax: int
-  edge_channels_list: List[int]
-  m_size: Tuple[int, ...]
-  cutoff: float
-  act_type: Literal['gate', 's2'] = 'gate'
-  to_grid_mat: jnp.ndarray | None = None
-  from_grid_mat: jnp.ndarray | None = None
-
-  @nn.compact
-  def __call__(
-    self,
-    x: jnp.ndarray,
-    x_edge: jnp.ndarray,
-    edge_distance: jnp.ndarray,
-    edge_index: jnp.ndarray,
-    wigner_and_M_mapping: jnp.ndarray,
-    wigner_and_M_mapping_inv: jnp.ndarray,
-    edge_envelope: jnp.ndarray,
-    node_offset: int = 0,
-  ) -> jnp.ndarray:
-    if self.act_type == 'gate':
-      extra_m0_output_channels = self.lmax * self.hidden_channels
-    else:
-      extra_m0_output_channels = self.hidden_channels
-
-    so2_conv_1 = SO2Convolution(
-      sphere_channels=2 * self.sphere_channels,
-      m_output_channels=self.hidden_channels,
-      lmax=self.lmax,
-      mmax=self.mmax,
-      m_size=self.m_size,
-      internal_weights=False,
-      edge_channels_list=self.edge_channels_list,
-      extra_m0_output_channels=extra_m0_output_channels,
-      name='so2_conv_1',
-    )
-
-    so2_conv_2 = SO2Convolution(
-      sphere_channels=self.hidden_channels,
-      m_output_channels=self.sphere_channels,
-      lmax=self.lmax,
-      mmax=self.mmax,
-      m_size=self.m_size,
-      internal_weights=True,
-      edge_channels_list=None,
-      extra_m0_output_channels=None,
-      name='so2_conv_2',
-    )
-
-    if self.act_type == 'gate':
-      act = GateActivation(
-        lmax=self.lmax,
-        mmax=self.mmax,
-        num_channels=self.hidden_channels,
-        m_prime=True,
-      )
-    else:
-      act = SeparableS2Activation(
-        lmax=self.lmax,
-        mmax=self.mmax,
-        to_grid_mat=self.to_grid_mat,
-        from_grid_mat=self.from_grid_mat,
-      )
-
-    x_source = x[edge_index[0]]
-    x_target = x[edge_index[1]]
-    x_message = jnp.concatenate([x_source, x_target], axis=2)
-    x_message = jnp.einsum('nmj,njc->nmc', wigner_and_M_mapping, x_message)
-
-    x_message, x_0_gating = so2_conv_1(x_message, x_edge)
-    x_message = act(x_0_gating, x_message)
-    x_message, _ = so2_conv_2(x_message, x_edge)
-    x_message = x_message * edge_envelope
-    x_message = jnp.einsum('njm,nmc->njc', wigner_and_M_mapping_inv, x_message)
-
-    target_indices = edge_index[1] - node_offset
-    new_embedding = jax.ops.segment_sum(
-      x_message,
-      target_indices,
-      num_segments=x.shape[0],
-    )
-
-    return new_embedding
-
-
-class GridAtomwise(nn.Module):
-  """Grid-based atomwise feed-forward module."""
-
-  sphere_channels: int
-  hidden_channels: int
-  lmax: int
-  mmax: int
-  to_grid_mat: jnp.ndarray
-  from_grid_mat: jnp.ndarray
-
-  @nn.compact
-  def __call__(self, x: jnp.ndarray) -> jnp.ndarray:
-    grid_mlp = nn.Sequential(
-      [
-        nn.Dense(self.hidden_channels, use_bias=False),
-        nn.silu,
-        nn.Dense(self.hidden_channels, use_bias=False),
-        nn.silu,
-        nn.Dense(self.sphere_channels, use_bias=False),
-      ]
-    )
-
-    x_grid = jnp.einsum('bai,zic->zbac', self.to_grid_mat, x)
-    x_grid = grid_mlp(x_grid)
-    x = jnp.einsum('bai,zbac->zic', self.from_grid_mat, x_grid)
-
-    return x
-
-
-class UMABlock(nn.Module):
-  """UMA transformer-like block."""
-
-  sphere_channels: int
-  hidden_channels: int
-  lmax: int
-  mmax: int
-  m_size: Tuple[int, ...]
-  edge_channels_list: List[int]
-  cutoff: float
-  norm_type: str = 'rms_norm_sh'
-  act_type: Literal['gate', 's2'] = 'gate'
-  ff_type: Literal['spectral', 'grid'] = 'grid'
-  to_grid_mat: jnp.ndarray | None = None
-  from_grid_mat: jnp.ndarray | None = None
-
-  @nn.compact
-  def __call__(
-    self,
-    x: jnp.ndarray,
-    x_edge: jnp.ndarray,
-    edge_distance: jnp.ndarray,
-    edge_index: jnp.ndarray,
-    wigner_and_M_mapping: jnp.ndarray,
-    wigner_and_M_mapping_inv: jnp.ndarray,
-    edge_envelope: jnp.ndarray,
-    sys_node_embedding: jnp.ndarray | None = None,
-    node_offset: int = 0,
-  ) -> jnp.ndarray:
-    # First residual block: Edgewise
-    x_res = x
-
-    norm_1 = get_normalization_layer(
-      self.norm_type,
-      lmax=self.lmax,
-      num_channels=self.sphere_channels,
-    )
-    x = norm_1(x)
-
-    if sys_node_embedding is not None:
-      x = x.at[:, 0, :].add(sys_node_embedding)
-
-    edgewise = Edgewise(
-      sphere_channels=self.sphere_channels,
-      hidden_channels=self.hidden_channels,
-      lmax=self.lmax,
-      mmax=self.mmax,
-      edge_channels_list=self.edge_channels_list,
-      m_size=self.m_size,
-      cutoff=self.cutoff,
-      act_type=self.act_type,
-      to_grid_mat=self.to_grid_mat,
-      from_grid_mat=self.from_grid_mat,
-      name='edge_wise',
-    )
-    x = edgewise(
-      x,
-      x_edge,
-      edge_distance,
-      edge_index,
-      wigner_and_M_mapping,
-      wigner_and_M_mapping_inv,
-      edge_envelope,
-      node_offset,
-    )
-    x = x + x_res
-
-    # Second residual block: Atomwise
-    x_res = x
-
-    norm_2 = get_normalization_layer(
-      self.norm_type,
-      lmax=self.lmax,
-      num_channels=self.sphere_channels,
-    )
-    x = norm_2(x)
-
-    atomwise = GridAtomwise(
-      sphere_channels=self.sphere_channels,
-      hidden_channels=self.hidden_channels,
-      lmax=self.lmax,
-      mmax=self.mmax,
-      to_grid_mat=self.to_grid_mat,
-      from_grid_mat=self.from_grid_mat,
-      name='atom_wise',
-    )
-    x = atomwise(x)
-    x = x + x_res
-
-    return x
-
-
-# -----------------------------------------------------------------------------
-# Main model: UMA Backbone
-# -----------------------------------------------------------------------------
-
-
-@dataclass
-class UMAConfig:
-  """Configuration for UMA model."""
-
-  max_num_elements: int = 100
-  sphere_channels: int = 128
-  lmax: int = 2
-  mmax: int = 2
-  num_layers: int = 2
-  hidden_channels: int = 128
-  cutoff: float = 5.0
-  edge_channels: int = 128
-  num_distance_basis: int = 512
-  norm_type: str = 'rms_norm_sh'
-  act_type: str = 'gate'
-  ff_type: str = 'grid'
-  grid_resolution: int | None = None
-  chg_spin_emb_type: str = 'pos_emb'
-  dataset_list: List[str] | None = field(default=None)
-  use_dataset_embedding: bool = True
-
-
-class UMABackbone(nn.Module):
-  """UMA backbone model."""
-
-  config: UMAConfig
-
-  def setup(self):
-    cfg = self.config
-    self.sph_feature_size = (cfg.lmax + 1) ** 2
-    self.mapping = create_coefficient_mapping(cfg.lmax, cfg.mmax)
-    self.so3_grid_lmax_lmax = create_so3_grid(
-      cfg.lmax, cfg.lmax, resolution=cfg.grid_resolution, rescale=True
-    )
-    self.so3_grid_lmax_mmax = create_so3_grid(
-      cfg.lmax, cfg.mmax, resolution=cfg.grid_resolution, rescale=True
-    )
-    self.Jd_list = load_jacobi_matrices_from_file(cfg.lmax)
-    self.coefficient_index = coefficient_idx(
-      self.so3_grid_lmax_lmax.mapping, cfg.lmax, cfg.mmax
-    )
-    self.edge_channels_list = [
-      cfg.num_distance_basis + 2 * cfg.edge_channels,
-      cfg.edge_channels,
-      cfg.edge_channels,
-    ]
-
-  @nn.compact
-  def __call__(
-    self,
-    positions: jnp.ndarray,
-    atomic_numbers: jnp.ndarray,
-    batch: jnp.ndarray,
-    edge_index: jnp.ndarray,
-    edge_distance_vec: jnp.ndarray,
-    charge: jnp.ndarray,
-    spin: jnp.ndarray,
-    dataset: List[str] | None = None,
-  ) -> Dict[str, jnp.ndarray]:
-    cfg = self.config
-    num_atoms = positions.shape[0]
-    num_edges = edge_index.shape[1]
-
-    # Embeddings
-    sphere_embedding = nn.Embed(
-      num_embeddings=cfg.max_num_elements,
-      features=cfg.sphere_channels,
-      name='sphere_embedding',
-    )
-    atom_emb = sphere_embedding(atomic_numbers)
-
-    charge_embedding = ChgSpinEmbedding(
-      embedding_type=cfg.chg_spin_emb_type,
-      embedding_target='charge',
-      embedding_size=cfg.sphere_channels,
-      trainable=False,
-      name='charge_embedding',
-    )
-    spin_embedding = ChgSpinEmbedding(
-      embedding_type=cfg.chg_spin_emb_type,
-      embedding_target='spin',
-      embedding_size=cfg.sphere_channels,
-      trainable=False,
-      name='spin_embedding',
-    )
-
-    chg_emb = charge_embedding(charge)
-    spin_emb = spin_embedding(spin)
-
-    if (
-      cfg.use_dataset_embedding
-      and cfg.dataset_list is not None
-      and dataset is not None
-    ):
-      dataset_embedding = DatasetEmbedding(
-        embedding_size=cfg.sphere_channels,
-        dataset_list=tuple(cfg.dataset_list),
-        trainable=False,
-        name='dataset_embedding',
-      )
-      dataset_emb = dataset_embedding(dataset)
-      csd_cat = jnp.concatenate([chg_emb, spin_emb, dataset_emb], axis=1)
-      mix_csd = nn.Dense(cfg.sphere_channels, name='mix_csd')
-      csd_mixed_emb = nn.silu(mix_csd(csd_cat))
-    else:
-      csd_cat = jnp.concatenate([chg_emb, spin_emb], axis=1)
-      mix_csd = nn.Dense(cfg.sphere_channels, name='mix_csd')
-      csd_mixed_emb = nn.silu(mix_csd(csd_cat))
-
-    # Edge distance embedding
-    edge_distance = jnp.linalg.norm(edge_distance_vec, axis=-1)
-    dist_scaled = edge_distance / cfg.cutoff
-
-    envelope = PolynomialEnvelope(exponent=5, name='envelope')
-    edge_envelope = envelope(dist_scaled).reshape(-1, 1, 1)
-
-    distance_expansion = GaussianSmearing(
-      start=0.0,
-      stop=cfg.cutoff,
-      num_gaussians=cfg.num_distance_basis,
-      basis_width_scalar=2.0,
-      name='distance_expansion',
-    )
-    edge_distance_embedding = distance_expansion(edge_distance)
-
-    source_embedding = nn.Embed(
-      num_embeddings=cfg.max_num_elements,
-      features=cfg.edge_channels,
-      embedding_init=initializers.uniform(scale=0.002),
-      name='source_embedding',
-    )
-    target_embedding = nn.Embed(
-      num_embeddings=cfg.max_num_elements,
-      features=cfg.edge_channels,
-      embedding_init=initializers.uniform(scale=0.002),
-      name='target_embedding',
-    )
-
-    source_emb = source_embedding(atomic_numbers[edge_index[0]])
-    target_emb = target_embedding(atomic_numbers[edge_index[1]])
-    source_emb = source_emb - 0.001
-    target_emb = target_emb - 0.001
-
-    x_edge = jnp.concatenate(
-      [edge_distance_embedding, source_emb, target_emb],
-      axis=1,
-    )
-
-    # Compute Wigner matrices
-    euler_angles = init_edge_rot_euler_angles(edge_distance_vec)
-    wigner = eulers_to_wigner(euler_angles, 0, cfg.lmax, self.Jd_list)
-    wigner_inv = jnp.transpose(wigner, (0, 2, 1))
-
-    if cfg.mmax != cfg.lmax:
-      wigner = wigner[:, self.coefficient_index, :][
-        :, :, self.coefficient_index
-      ]
-      wigner_inv = wigner_inv[:, self.coefficient_index, :][
-        :, :, self.coefficient_index
-      ]
-
-    to_m = self.mapping.to_m
-    wigner_and_M_mapping = jnp.einsum('mk,nkj->nmj', to_m, wigner)
-    wigner_and_M_mapping_inv = jnp.einsum('njk,mk->njm', wigner_inv, to_m)
-
-    # Initialize node embeddings
-    x_message = jnp.zeros(
-      (num_atoms, self.sph_feature_size, cfg.sphere_channels),
-      dtype=positions.dtype,
-    )
-    x_message = x_message.at[:, 0, :].set(atom_emb)
-
-    sys_node_embedding = csd_mixed_emb[batch]
-    x_message = x_message.at[:, 0, :].add(sys_node_embedding)
-
-    # Edge degree embedding
-    edge_degree_embedding = EdgeDegreeEmbedding(
-      sphere_channels=cfg.sphere_channels,
-      lmax=cfg.lmax,
-      mmax=cfg.mmax,
-      edge_channels_list=self.edge_channels_list,
-      rescale_factor=5.0,
-      m_size=self.mapping.m_size,
-      name='edge_degree_embedding',
-    )
-    x_message = edge_degree_embedding(
-      x_message,
-      x_edge,
-      edge_index,
-      wigner_and_M_mapping_inv,
-      edge_envelope,
-      node_offset=0,
-    )
-
-    # Message passing blocks
-    for i in range(cfg.num_layers):
-      block = UMABlock(
-        sphere_channels=cfg.sphere_channels,
-        hidden_channels=cfg.hidden_channels,
-        lmax=cfg.lmax,
-        mmax=cfg.mmax,
-        m_size=self.mapping.m_size,
-        edge_channels_list=self.edge_channels_list,
-        cutoff=cfg.cutoff,
-        norm_type=cfg.norm_type,
-        act_type=cfg.act_type,
-        ff_type=cfg.ff_type,
-        to_grid_mat=self.so3_grid_lmax_lmax.to_grid_mat,
-        from_grid_mat=self.so3_grid_lmax_lmax.from_grid_mat,
-        name=f'blocks_{i}',
-      )
-      x_message = block(
-        x_message,
-        x_edge,
-        edge_distance,
-        edge_index,
-        wigner_and_M_mapping,
-        wigner_and_M_mapping_inv,
-        edge_envelope,
-        sys_node_embedding=sys_node_embedding,
-        node_offset=0,
-      )
-
-    # Final normalization
-    norm = get_normalization_layer(
-      cfg.norm_type,
-      lmax=cfg.lmax,
-      num_channels=cfg.sphere_channels,
-    )
-    x_message = norm(x_message)
-
-    return {
-      'node_embedding': x_message,
-      'batch': batch,
-    }
-
-
-def test_pytorch_model():
-  """Test PyTorch UMA model if available."""
-  try:
-    import torch
-
-    # Add fairchem to path
-    fairchem_path = os.path.join(
-      os.path.dirname(__file__), '..', '..', '..', '..', '..', 'fairchem', 'src'
-    )
-    if os.path.exists(fairchem_path):
-      sys.path.insert(0, fairchem_path)
-
-    from fairchem.core.models.uma.escn_md import eSCNMDBackbone
-
-    return True, eSCNMDBackbone
-  except ImportError as e:
-    print(f'PyTorch model not available: {e}')
-    return False, None
-
-
-def create_test_data(num_atoms=10, num_systems=2, seed=42):
-  """Create synthetic test data for both frameworks."""
-  np.random.seed(seed)
-
-  # Positions
-  positions = np.random.randn(num_atoms, 3).astype(np.float32) * 2.0
-
-  # Atomic numbers (H, C, N, O)
-  atomic_numbers = np.random.choice([1, 6, 7, 8], size=num_atoms).astype(
-    np.int32
-  )
-
-  # Batch indices
-  atoms_per_system = num_atoms // num_systems
-  batch = np.repeat(np.arange(num_systems), atoms_per_system).astype(np.int32)
-  if len(batch) < num_atoms:
-    batch = np.concatenate(
-      [batch, np.full(num_atoms - len(batch), num_systems - 1)]
-    ).astype(np.int32)
-
-  # Build edges (within cutoff)
-  cutoff = 5.0
-  edge_src = []
-  edge_dst = []
-  for i in range(num_atoms):
-    for j in range(num_atoms):
-      if i != j:
-        dist = np.linalg.norm(positions[i] - positions[j])
-        if dist < cutoff:
-          edge_src.append(i)
-          edge_dst.append(j)
-
-  edge_index = np.array([edge_src, edge_dst], dtype=np.int32)
-  edge_distance_vec = (
-    positions[edge_index[0]] - positions[edge_index[1]]
-  ).astype(np.float32)
-
-  # System-level properties
-  charge = np.zeros(num_systems, dtype=np.float32)
-  spin = np.zeros(num_systems, dtype=np.float32)
-  dataset = ['omat'] * num_systems
-  natoms = np.array([atoms_per_system] * num_systems, dtype=np.int32)
-
-  return {
-    'positions': positions,
-    'atomic_numbers': atomic_numbers,
-    'batch': batch,
-    'edge_index': edge_index,
-    'edge_distance_vec': edge_distance_vec,
-    'charge': charge,
-    'spin': spin,
-    'dataset': dataset,
-    'natoms': natoms,
-  }
-
-
-def test_jax_forward_pass():
-  """Test JAX UMA forward pass."""
-  config = UMAConfig(
-    max_num_elements=100,
-    sphere_channels=64,  # Smaller for testing
-    lmax=2,
-    mmax=2,
-    num_layers=2,
-    hidden_channels=64,
-    cutoff=5.0,
-    edge_channels=64,
-    num_distance_basis=128,  # Smaller for testing
-    norm_type='rms_norm_sh',
-    act_type='gate',
-    ff_type='grid',
-    chg_spin_emb_type='pos_emb',
-    dataset_list=['oc20', 'omol', 'omat', 'odac', 'omc'],
-    use_dataset_embedding=True,
-  )
-
-  # Create model
-  model = UMABackbone(config=config)
-
-  # Create test data
-  data = create_test_data(num_atoms=10, num_systems=2)
-
-  # Convert to JAX arrays
-  positions = jnp.array(data['positions'])
-  atomic_numbers = jnp.array(data['atomic_numbers'])
-  batch = jnp.array(data['batch'])
-  edge_index = jnp.array(data['edge_index'])
-  edge_distance_vec = jnp.array(data['edge_distance_vec'])
-  charge = jnp.array(data['charge'])
-  spin = jnp.array(data['spin'])
-  from jax_md._nn.uma.nn.embedding import dataset_names_to_indices
-
-  dataset_idx = dataset_names_to_indices(data['dataset'], config.dataset_list)
-
-  # Initialize parameters
-  key = jax.random.PRNGKey(0)
-  params = model.init(
-    key,
-    positions,
-    atomic_numbers,
-    batch,
-    edge_index,
-    edge_distance_vec,
-    charge,
-    spin,
-    dataset_idx,
-  )
-
-  # Forward pass
-  output = model.apply(
-    params,
-    positions,
-    atomic_numbers,
-    batch,
-    edge_index,
-    edge_distance_vec,
-    charge,
-    spin,
-    dataset_idx,
-  )
-
-  node_embedding = output['node_embedding']
-  assert node_embedding.shape == (10, 9, 64)
-  assert jnp.all(jnp.isfinite(node_embedding))
-
-
-def test_pytorch_forward_pass():
-  """Test PyTorch UMA forward pass if available."""
-  available, eSCNMDBackbone = test_pytorch_model()
-  if not available:
-    return False, None, None
-
-  import torch
-
-  # Create PyTorch model with same config
-  model = eSCNMDBackbone(
-    max_num_elements=100,
-    sphere_channels=64,
-    lmax=2,
-    mmax=2,
-    num_layers=2,
-    hidden_channels=64,
-    cutoff=5.0,
-    edge_channels=64,
-    num_distance_basis=128,
-    norm_type='rms_norm_sh',
-    act_type='gate',
-    ff_type='grid',
-    chg_spin_emb_type='pos_emb',
-    dataset_list=['oc20', 'omol', 'omat', 'odac', 'omc'],
-    use_dataset_embedding=True,
-    otf_graph=False,
-  )
-  model.eval()
-
-  # Create test data
-  data = create_test_data(num_atoms=10, num_systems=2)
-
-  # FairChem expects a dict-like with .get(key, default=...) keyword support
-  class AtomicDataDict(dict):
-    def __getattr__(self, k):
-      try:
-        return self[k]
-      except KeyError:
-        raise AttributeError(k)
-
-    def __setattr__(self, k, v):
-      self[k] = v
-
-    def get(self, k, *a, **kw):
-      return super().get(k, kw.get('default', a[0] if a else None))
-
-  data_dict = AtomicDataDict(
-    pos=torch.tensor(data['positions']),
-    atomic_numbers=torch.tensor(data['atomic_numbers']).long(),
-    atomic_numbers_full=torch.tensor(data['atomic_numbers']).long(),
-    batch=torch.tensor(data['batch']).long(),
-    batch_full=torch.tensor(data['batch']).long(),
-    edge_index=torch.tensor(data['edge_index']).long(),
-    cell_offsets=torch.zeros(data['edge_index'].shape[1], 3),
-    cell=torch.eye(3).unsqueeze(0).expand(2, 3, 3),
-    charge=torch.tensor(data['charge']),
-    spin=torch.tensor(data['spin']),
-    dataset=data['dataset'],
-    natoms=torch.tensor(data['natoms']),
-    nedges=torch.tensor(
-      [data['edge_index'].shape[1] // 2, data['edge_index'].shape[1] // 2]
-    ),
-  )
-
-  # Forward pass
-  with torch.no_grad():
-    output = model(data_dict)
-
-  node_embedding = output['node_embedding']
-  assert node_embedding.shape == (10, 9, 64)
-  assert torch.all(torch.isfinite(node_embedding))
-
-
-if __name__ == '__main__':
-  test_jax_forward_pass()
-  test_pytorch_forward_pass()
